requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
idna==2.8
multidict==4.5.2
PyYAML==5.1.2
six==1.12.0
vcrpy==2.1.0
wrapt==1.11.2
yarl==1.3.0
